Added ability to set frame to render on single image render tab. Added support for specifying absolute paths in the ClientSettings and ServerSettings files. Added .json extension to ClientSettings and ServerSettings files for ease of use.

+++ LogicReinc.BlendFarm.Server/Program.cs
@@ -62,9 +62,12 @@ namespace LogicReinc.BlendFarm.Server
         /// </summary>
         public static void CleanupOldSessions()
         {
+            if (!ServerSettings.Instance.DeleteBlenderFilesOnExit) return;
+
             try
             {
-                string path = SystemInfo.RelativeToApplicationDirectory(ServerSettings.Instance.BlenderFiles);
+                //string path = SystemInfo.RelativeToApplicationDirectory(ServerSettings.Instance.BlenderFiles);
+                string path = ServerSettings.Instance.GetBlenderFilesPath();
                 if (Directory.Exists(path))
                     Directory.Delete(path, true);
             }
